% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_update.R
\name{add_to_sysdata}
\alias{add_to_sysdata}
\title{add_to_sysdata}
\usage{
add_to_sysdata(..., silent = F, overwrite = F)
}
\arguments{
\item{silent}{logical for messages}
}
\value{
message
}
\description{
Load sysdata.rda if it exists and add objects in \code{...} to it.
}
