% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_update.R
\name{bump_version}
\alias{bump_version}
\title{bump_version}
\usage{
bump_version(which = "dev")
}
\arguments{
\item{which}{A string specifying which level to increment, one of: "major",
"minor", "patch", "dev". If \code{NULL}, user can choose interactively.}
}
\value{
bump
}
\description{
bump version
}
