% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_run_dev.R
\name{copy_golem_to_wd}
\alias{copy_golem_to_wd}
\title{copy_golem_to_wd}
\usage{
copy_golem_to_wd(overwrite, silent)
}
\arguments{
\item{overwrite}{logical for overwrite}

\item{silent}{logical for messages}
}
\value{
files being copied if needed/wanted
}
\description{
copy minumim golem files to working directory
}
