% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_run_dev.R
\name{delete_combined_tests}
\alias{delete_combined_tests}
\title{delete_combined_tests}
\usage{
delete_combined_tests()
}
\value{
message
}
\description{
delete tests.R file in dev for when you pull a new update from github
}
