% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_update.R
\name{delete_combined_vignettes}
\alias{delete_combined_vignettes}
\title{delete_combined_vignettes}
\usage{
delete_combined_vignettes()
}
\value{
message
}
\description{
delete tests.R file in dev for when you pull a new update from github
}
