% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_update.R
\name{dev_pull_and_update}
\alias{dev_pull_and_update}
\title{dev_pull_and_update}
\usage{
dev_pull_and_update(silent = FALSE)
}
\arguments{
\item{silent}{logical for messages}
}
\value{
message
}
\description{
Update package from combined.R by documenting and combining files again
}
