% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_update.R
\name{dev_update_commit_push}
\alias{dev_update_commit_push}
\title{dev_update_commit_push}
\usage{
dev_update_commit_push(
  silent = FALSE,
  is_production = FALSE,
  overwrite = FALSE,
  message = "dev"
)
}
\arguments{
\item{silent}{logical for messages}
}
\value{
message
}
\description{
Update package from combined.R by documenting and combining files again
}
