% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_split_combine.R
\name{fast_commit}
\alias{fast_commit}
\title{Fast Commit}
\usage{
fast_commit(comment = "dev")
}
\arguments{
\item{comment}{character string for commit message}
}
\value{
commited git
}
\description{
commit for git with one function
}
