% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_split_combine.R
\name{find_imported_functions}
\alias{find_imported_functions}
\title{find_imported_functions}
\usage{
find_imported_functions(file_path = "dev/combined.R")
}
\value{
functions
}
\description{
find_imported_functions
}
