% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_other.R
\name{get_external_functions}
\alias{get_external_functions}
\title{get_external_functions}
\usage{
get_external_functions(pkg)
}
\arguments{
\item{pkg}{character vector of pkg}
}
\description{
get_external_functions
}
