% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_split_combine.R
\name{pkg_combine_R_files}
\alias{pkg_combine_R_files}
\title{pkg_combine_R_files}
\usage{
pkg_combine_R_files(
  pkg,
  destination_dir = getwd(),
  filename = "combined.R",
  header_symbol = "=",
  max_new_lines = 0,
  new_lines = character(0)
)
}
\arguments{
\item{pkg}{package name as character string}
}
\value{
message
}
\description{
pkg_combine_R_files
}
