% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_nav.R
\name{pkg_dev_nav}
\alias{pkg_dev_nav}
\title{pkg_dev_nav}
\usage{
pkg_dev_nav(name = NULL)
}
\arguments{
\item{name}{character string for the file name to be opened}
}
\value{
file opens in RStudio
}
\description{
Navigate to dev file in RStudio
}
