% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_nav.R
\name{pkg_dev_nav_combined}
\alias{pkg_dev_nav_combined}
\title{pkg_dev_nav_combined}
\usage{
pkg_dev_nav_combined(pane = T)
}
\value{
file opens in RStudio
}
\description{
Navigate to combined.R dev file in RStudio
}
