% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_other.R
\name{pkg_net_node_edges}
\alias{pkg_net_node_edges}
\title{pkg_net_node_edges}
\usage{
pkg_net_node_edges(
  pkg_name,
  show_internal = FALSE,
  physics = TRUE,
  arrows = "to"
)
}
\description{
pkg_net_node_edges
}
