% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_other.R
\name{remove_trailing_whitespace}
\alias{remove_trailing_whitespace}
\title{remove_trailing_whitespace}
\usage{
remove_trailing_whitespace(file_path = "dev/combined.R")
}
\description{
remove_trailing_whitespace
}
