% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_split_combine.R
\name{split_R_files}
\alias{split_R_files}
\title{split_R_files}
\usage{
split_R_files(
  source_dir = file.path(getwd(), "dev"),
  destination_dir = file.path(getwd(), "R"),
  file_name = "combined",
  file_ext = ".R",
  header_symbol = "=",
  new_lines = character(0),
  silent = FALSE
)
}
\arguments{
\item{source_dir}{a file path for your source (such as R folder)}

\item{destination_dir}{a file path for your destination (such as dev folder)}

\item{file_name}{a file name (ends in .R)}

\item{header_symbol}{single character for your header separator in combined file}

\item{new_lines}{character vector for new lines}
}
\value{
message
}
\description{
split_R_files
}
