% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_other.R
\name{check_namespace_conflicts}
\alias{check_namespace_conflicts}
\title{check_namespace_conflicts}
\usage{
check_namespace_conflicts(
  pkgs,
  ignores = c("\%>\%", "pkg_date", "pkg_version", "pkg_name", ".__NAMESPACE__.",
    ".__S3MethodsTable__.", ".packageName")
)
}
\arguments{
\item{pkgs}{character vector of pkgs}
}
\description{
check_namespace_conflicts
}
