% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_update.R
\name{copy_golem_to_wd}
\alias{copy_golem_to_wd}
\title{copy_golem_to_wd}
\usage{
copy_golem_to_wd(overwrite = FALSE, silent = TRUE)
}
\arguments{
\item{overwrite}{logical for overwrite}

\item{silent}{logical for messages}
}
\value{
files being copied if needed/wanted
}
\description{
copy minimum golem files to working directory
}
