% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_update.R
\name{dev_update}
\alias{dev_update}
\title{Update Dev}
\usage{
dev_update(
  silent = FALSE,
  use_internal_pkg = TRUE,
  is_production = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{silent}{logical for messages}
}
\value{
message
}
\description{
Update package from combined.R by documenting and combining files again
}
