% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_run_dev.R
\name{dev_update}
\alias{dev_update}
\title{Update Dev}
\usage{
dev_update(silent = F)
}
\arguments{
\item{silent}{logical for messages}
}
\value{
message
}
\description{
Update package from combined.R by documenting and combining files again
}
