% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_run_dev.R
\name{fast_commit}
\alias{fast_commit}
\title{Fast Commit}
\usage{
fast_commit(message = "dev", push = F)
}
\arguments{
\item{message}{Message to use for first commit.}

\item{push}{If \code{TRUE}, also attempts to push the commits to the remote
branch.}
}
\value{
commited git
}
\description{
commit for git with one function
}
