% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_run_dev.R
\name{fast_commit}
\alias{fast_commit}
\title{Fast Commit}
\usage{
fast_commit(message = "dev", push = F, bump_version = F, which = "dev")
}
\arguments{
\item{comment}{character string for commit message}
}
\value{
commited git
}
\description{
commit for git with one function
}
