% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_split_combine.R
\name{pkg_combine_R_files_launch}
\alias{pkg_combine_R_files_launch}
\title{pkg_combine_R_files_launch}
\usage{
pkg_combine_R_files_launch(
  pkg,
  destination_dir = tempdir(),
  filename_type = "name__version",
  header_symbol = "=",
  max_new_lines = 0,
  new_lines = character(0),
  overwrite = F
)
}
\arguments{
\item{filename_type}{character string of type filename: "name__version" or "name"}

\item{pkgs}{package name(s) as character string}
}
\value{
message
}
\description{
pkg_combine_R_files_launch
}
