% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RosyDev-package.R
\docType{package}
\name{RosyDev-package}
\alias{RosyDev}
\alias{RosyDev-package}
\title{RosyDev: Rosy Dev}
\description{
Rosy development for your R package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/brandonerose/RosyDev}
  \item Report bugs at \url{https://github.com/brandonerose/RosyDev/issues}
}

}
\author{
\strong{Maintainer}: Brandon Rose \email{thecodingdocs@gmail.com}

}
\keyword{internal}
