% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_run_dev.R
\name{setup_Rosydev}
\alias{setup_Rosydev}
\title{Package System}
\usage{
setup_Rosydev(silent = F, launch_files = T)
}
\arguments{
\item{silent}{logical for messages}

\item{launch_files}{logical for launching files}
}
\value{
path
}
\description{
Find the file system of a package
}
