% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_run_dev.R
\name{setup_RosyDev}
\alias{setup_RosyDev}
\title{Package System}
\usage{
setup_RosyDev(silent = F, launch_files = T, overwrite = F, use_golem = F)
}
\arguments{
\item{silent}{logical for messages}

\item{launch_files}{logical for launching files}

\item{overwrite}{logical for overwrite}

\item{use_golem}{logical for using golem}
}
\value{
path
}
\description{
Find the file system of a package
}
